# tests/contract/snapshots/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ErrorCode": {
        "enum": [
          "REQUEST_VALIDATION_FAILED",
          "INVALID_CURSOR",
          "INVALID_IDEMPOTENCY_KEY",
          "IDEMPOTENCY_KEY_REUSED",
          "CAPABILITY_NOT_ENABLED",
          "ANSWER_POLICY_NOT_SUPPORTED",
          "CHAT_SESSION_BUSY",
          "CHAT_CONTEXT_INVALID",
          "CHAT_REVISION_MISMATCH",
          "CHAT_PROVIDER_UNAVAILABLE",
          "CHAT_RESPONSE_INVALID",
          "CHAT_PIPELINE_DEADLINE_EXCEEDED",
          "CHAT_PIPELINE_STEP_FAILED",
          "CHAT_PERSISTENCE_FAILED",
          "CHAT_STALE_WORKER",
          "CHAT_WORKER_STOPPED",
          "CHAT_RUN_CANCELLED",
          "CHAT_SSE_CONNECTION_LIMIT_EXCEEDED",
          "IDENTITY_OVERRIDE_NOT_ALLOWED",
          "ACCESS_DENIED",
          "RESOURCE_NOT_FOUND",
          "METHOD_NOT_ALLOWED",
          "HTTP_ERROR",
          "INTERNAL_SERVER_ERROR",
          "RESOURCE_NAME_CONFLICT",
          "RESOURCE_STATE_CONFLICT",
          "FILE_NAME_INVALID",
          "FILE_MEDIA_TYPE_UNSUPPORTED",
          "FILE_MEDIA_TYPE_MISMATCH",
          "FILE_TOO_LARGE",
          "FILE_INVALID_UTF8",
          "FILE_LINE_LIMIT_EXCEEDED",
          "FILE_STRUCTURE_LIMIT_EXCEEDED",
          "FILE_CONTENT_INVALID",
          "FILE_ARCHIVE_LIMIT_EXCEEDED",
          "MARKDOWN_BUNDLE_INVALID",
          "MARKDOWN_MEDIA_UNRESOLVED",
          "MARKDOWN_MEDIA_UNSUPPORTED",
          "MARKDOWN_MEDIA_FETCH_FAILED",
          "PARSER_NOT_CONFIGURED",
          "PARSER_CHUNK_LIMIT_EXCEEDED",
          "PARSER_RESOURCE_LIMIT",
          "PARSER_CRASHED",
          "PARSER_OUTPUT_INVALID",
          "INDEX_TARGET_INVALID",
          "INDEX_REVISION_INCOMPATIBLE",
          "INDEX_CHUNK_PLAN_MISMATCH",
          "SEMANTIC_CHUNKING_FAILED",
          "SOURCE_NOT_AVAILABLE",
          "SOURCE_FILE_MISSING",
          "SOURCE_FILE_INTEGRITY",
          "EMBEDDING_SPACE_MISMATCH",
          "EMBEDDING_PROVIDER_UNAVAILABLE",
          "EMBEDDING_RESPONSE_INVALID",
          "LOCAL_RERANKER_UNAVAILABLE",
          "RETRIEVAL_DEADLINE_EXCEEDED",
          "GRAPH_NOT_READY",
          "GRAPH_CONFIG_INVALID",
          "GRAPH_PROVIDER_UNAVAILABLE",
          "GRAPH_PROTOCOL_INVALID",
          "GRAPH_RESOURCE_LIMIT",
          "GRAPH_BUILD_FAILED",
          "INDEX_PERSISTENCE_FAILED",
          "INDEX_INCOMPLETE",
          "INDEXING_STALE_WORKER",
          "INDEXING_WORKER_STOPPED"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
...
      "FieldViolation": {
        "additionalProperties": false,
        "properties": {
          "error_type": {
            "title": "Error Type",
            "type": "string"
          },
          "location": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "location",
          "message",
          "error_type"
        ],
        "title": "FieldViolation",
        "type": "object"
      },
...
      "IndexingErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "additionalProperties": true,
            "title": "Detail",
            "type": "object"
          }
        },
        "required": [
          "code",
          "detail"
        ],
        "title": "IndexingErrorResponse",
        "type": "object"
      },
      "IndexingJobPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IndexingJobResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "IndexingJobPage",
        "type": "object"
      },
      "IndexingJobResponse": {
        "additionalProperties": false,
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "build_status": {
            "enum": [
              "queued",
              "processing",
              "ready",
              "failed"
            ],
            "title": "Build Status",
            "type": "string"
          },
          "can_retry": {
            "title": "Can Retry",
            "type": "boolean"
          },
          "claimed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "document_version_id": {
            "format": "uuid",
            "title": "Document Version Id",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IndexingErrorResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "heartbeat_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heartbeat At"
          },
          "index_revision_id": {
            "format": "uuid",
            "title": "Index Revision Id",
            "type": "string"
          },
          "indexed_document_version_id": {
            "format": "uuid",
            "title": "Indexed Document Version Id",
            "type": "string"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "kb_id": {
            "format": "uuid",
            "title": "Kb Id",
            "type": "string"
          },
          "next_attempt_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Attempt At"
          },
          "phase": {
            "title": "Phase",
            "type": "string"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IndexingProgressResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "serving_status": {
            "enum": [
              "candidate",
              "serving",
              "retired"
            ],
            "title": "Serving Status",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "kb_id",
          "document_id",
          "document_version_id",
          "indexed_document_version_id",
          "index_revision_id",
          "status",
          "phase",
          "progress",
          "attempt",
          "build_status",
          "serving_status",
          "claimed_at",
          "heartbeat_at",
          "next_attempt_at",
          "error",
          "can_retry",
          "created_at",
          "updated_at"
        ],
        "title": "IndexingJobResponse",
        "type": "object"
      },
      "IndexingProgressResponse": {
        "additionalProperties": false,
        "properties": {
          "child_peak_rss_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Child Peak Rss Bytes"
          },
          "completed_pages": {
            "title": "Completed Pages",
            "type": "integer"
          },
          "elapsed_ms": {
            "title": "Elapsed Ms",
            "type": "integer"
          },
          "ocr_pages": {
            "title": "Ocr Pages",
            "type": "integer"
          },
          "ocr_regions": {
            "title": "Ocr Regions",
            "type": "integer"
          },
          "page_from": {
            "title": "Page From",
            "type": "integer"
          },
          "page_to": {
            "title": "Page To",
            "type": "integer"
          },
          "schema_version": {
            "const": "pdf_parsing_progress_v1",
            "title": "Schema Version",
            "type": "string"
          },
          "segment_count": {
            "title": "Segment Count",
            "type": "integer"
          },
          "segment_number": {
            "title": "Segment Number",
            "type": "integer"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "stage_pages": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Stage Pages",
            "type": "object"
          },
          "table_candidates": {
            "title": "Table Candidates",
            "type": "integer"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "schema_version",
          "stage",
          "total_pages",
          "completed_pages",
          "segment_number",
          "segment_count",
          "page_from",
          "page_to",
          "stage_pages",
          "ocr_pages",
          "ocr_regions",
          "table_candidates",
          "elapsed_ms"
        ],
        "title": "IndexingProgressResponse",
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldViolation"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "trace_id",
          "retryable"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Enterprise Knowledge Base API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/indexing-jobs/{job_id}": {
      "get": {
        "operationId": "get_indexing_job_api_v1_indexing_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexingJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Get Indexing Job",
        "tags": [
          "indexing"
        ]
      }
    },
    "/api/v1/indexing-jobs/{job_id}/retry": {
      "post": {
        "operationId": "retry_indexing_job_api_v1_indexing_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "description": "Required UUID scoped by principal, client, and endpoint.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "description": "Required UUID scoped by principal, client, and endpoint.",
              "format": "uuid",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexingJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Idempotency or resource conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Retry Indexing Job",
        "tags": [
          "indexing"
        ]
      }
    },
...
    "/api/v1/knowledge-bases/{kb_id}/indexing-jobs": {
      "get": {
        "operationId": "list_indexing_jobs_api_v1_knowledge_bases__kb_id__indexing_jobs_get",
        "parameters": [
          {
            "in": "path",
            "name": "kb_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Kb Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2048,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexingJobPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid request or cursor"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "List Indexing Jobs",
        "tags": [
          "indexing"
        ]
      }
    },
...
  }
}